Created highlight property
