+++ Source/Basic-Conditions-And-Behaviors/Runtime/Properties/Highlighting/IOutlineHighlightProperty.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60a990716c43ab547b34c7f02f364305
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
