+++ Source/Basic-Conditions-And-Behaviors/Runtime/Properties/Highlighting/OutlineHighlightProperty.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fabd57f28c4a5ed41be949b8c4f0b3ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: f131347375d03214588a261e8d223ad7, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
